requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.5
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.12.0
django-on-heroku==1.1.2
django-utils-six==2.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.1
redis==3.5.3
redis-py-cluster==2.1.3
sqlparse==0.4.2
whitenoise==6.2.0
python-dotenv